add hw13 (hibernate)

+++ src/main/java/ua/com/alevel/hw2/service/invoiceservice/InvoiceService.java
@@ -56,7 +56,7 @@ public final class InvoiceService {
         invoiceDaoJPA.updateDate(id, date);
     }
 
-    public Map<Double, Integer> groupingBySum() {
+    public Map<Double, Long> groupingBySum() {
         return invoiceDaoJPA.groupingBySum();
     }
 }
